+++ activerecord/test/cases/relation/mutation_test.rb
@@ -156,8 +156,7 @@ module ActiveRecord
 
     test 'none!' do
       assert relation.none!.equal?(relation)
-      assert_equal [NullRelation], relation.extending_values
-      assert relation.is_a?(NullRelation)
+      assert relation.null?
     end
 
     test 'distinct!' do
